+++ sources/s3_bucket/s3_bucket_source.go
@@ -21,6 +21,7 @@ import (
 	"github.com/turbot/pipe-fittings/v2/filter"
 	"github.com/turbot/tailpipe-plugin-aws/config"
 	"github.com/turbot/tailpipe-plugin-sdk/artifact_source"
+	"github.com/turbot/tailpipe-plugin-sdk/context_values"
 	"github.com/turbot/tailpipe-plugin-sdk/row_source"
 	"github.com/turbot/tailpipe-plugin-sdk/types"
 )
@@ -34,8 +35,7 @@ const (
 type AwsS3BucketSource struct {
 	artifact_source.ArtifactSourceImpl[*AwsS3BucketSourceConfig, *config.AwsConnection]
 
-	client    *s3.Client
-	errorList []error
+	client *s3.Client
 }
 
 func (s *AwsS3BucketSource) Init(ctx context.Context, params *row_source.RowSourceParams, opts ...row_source.RowSourceOption) error {
@@ -49,11 +49,10 @@ func (s *AwsS3BucketSource) Init(ctx context.Context, params *row_source.RowSour
 	// initialize client
 	client, err := s.getClient(ctx)
 	if err != nil {
+		slog.Error("Error getting S3 client", "error", err)
 		return err
 	}
 	s.client = client
-
-	s.errorList = []error{}
 
 	slog.Info("Initialized AwsS3BucketSource", "bucket", s.Config.Bucket, "layout", s.Config.FileLayout)
 
@@ -89,6 +88,8 @@ func (s *AwsS3BucketSource) DiscoverArtifacts(ctx context.Context) error {
 	// add any patterns defined in config
 	err := g.AddPatterns(s.Config.GetPatterns())
 	if err != nil {
+		// fatal error - log and return
+		slog.Error("error adding grok patterns", "error", err)
 		return fmt.Errorf("error adding grok patterns: %v", err)
 	}
 
@@ -104,14 +105,13 @@ func (s *AwsS3BucketSource) DiscoverArtifacts(ctx context.Context) error {
 		optionalLayouts = newOptionalLayouts
 	}
 
+	// walkS3 should only return fatal errors
 	err = s.walkS3(ctx, s.Config.Bucket, prefix, optionalLayouts, filterMap, g)
 	if err != nil {
-		s.errorList = append(s.errorList, fmt.Errorf("error discovering artifacts in S3 bucket %s, %w", s.Config.Bucket, err))
+		slog.Error("error walking S3 bucket", "bucket", s.Config.Bucket, "error", err)
+		return fmt.Errorf("%s: %s", s.Config.Bucket, err.Error())
 	}
 
-	if len(s.errorList) > 0 {
-		return errors.Join(s.errorList...)
-	}
 	return nil
 }
 
@@ -122,7 +122,8 @@ func (s *AwsS3BucketSource) DownloadArtifact(ctx context.Context, info *types.Ar
 		Key:    &info.Name,
 	})
 	if err != nil {
-		return fmt.Errorf("failed to download artifact, %w", err)
+		slog.Error("failed to download artifact", "bucket", s.Config.Bucket, "key", info.Name, "error", err)
+		return fmt.Errorf("%s: failed to download artifact from %s", info.Name, s.Config.Bucket)
 	}
 	defer getObjectOutput.Body.Close()
 
@@ -131,28 +132,31 @@ func (s *AwsS3BucketSource) DownloadArtifact(ctx context.Context, info *types.Ar
 
 	// copy the object data to a temp file
 	localFilePath := path.Join(s.TempDir, info.Name)
+	localFileDir := filepath.Dir(localFilePath)
+
 	// ensure the directory exists of the file to write to
-	if err := os.MkdirAll(filepath.Dir(localFilePath), 0755); err != nil {
-		return fmt.Errorf("failed to create directory for file, %w", err)
+	if err := os.MkdirAll(localFileDir, 0755); err != nil {
+		slog.Error("failed to create directory", "bucket", s.Config.Bucket, "key", info.Name, "dir", localFileDir, "error", err)
+		return fmt.Errorf("%s: failed to download artifact from %s", info.Name, s.Config.Bucket)
 	}
 
 	// Create a local file to write the data to
 	outFile, err := os.Create(localFilePath)
 	if err != nil {
-		return fmt.Errorf("failed to create file, %w", err)
+		slog.Error("failed to create file", "bucket", s.Config.Bucket, "key", info.Name, "file", outFile, "error", err)
+		return fmt.Errorf("%s: failed to download artifact from %s", info.Name, s.Config.Bucket)
 	}
 	defer outFile.Close()
 
 	// Write the data to the local file
 	_, err = io.Copy(outFile, getObjectOutput.Body)
 	if err != nil {
-		return fmt.Errorf("failed to write data to file, %w", err)
+		slog.Error("failed to write file content", "bucket", s.Config.Bucket, "key", info.Name, "file", outFile, "error", err)
+		return fmt.Errorf("%s: failed to download artifact from %s", info.Name, s.Config.Bucket)
 	}
 
 	// notify observers of the downloaded artifact
-	downloadInfo := types.NewDownloadedArtifactInfo(info, localFilePath, size)
-
-	return s.OnArtifactDownloaded(ctx, downloadInfo)
+	return s.OnArtifactDownloaded(ctx, types.NewDownloadedArtifactInfo(info, localFilePath, size))
 }
 
 func (s *AwsS3BucketSource) getClient(ctx context.Context) (*s3.Client, error) {
@@ -180,6 +184,11 @@ func (s *AwsS3BucketSource) getClient(ctx context.Context) (*s3.Client, error) {
 }
 
 func (s *AwsS3BucketSource) walkS3(ctx context.Context, bucket string, prefix string, layouts []string, filterMap map[string]*filter.SqlFilter, g *grok.Grok) error {
+	executionId, err := context_values.ExecutionIdFromContext(ctx)
+	if err != nil {
+		return err
+	}
+
 	paginator := s3.NewListObjectsV2Paginator(s.client, &s3.ListObjectsV2Input{
 		Bucket:    aws.String(bucket),
 		Prefix:    aws.String(prefix),
@@ -189,29 +198,48 @@ func (s *AwsS3BucketSource) walkS3(ctx context.Context, bucket string, prefix st
 	for paginator.HasMorePages() {
 		page, err := paginator.NextPage(ctx)
 		if err != nil {
+			// fatal error - log and return
+			slog.Error("error getting next page", "bucket", bucket, "prefix", prefix, "error", err)
 			return fmt.Errorf("error getting next page, %w", err)
 		}
 
 		// Directories
 		for _, dir := range page.CommonPrefixes {
-			err = s.WalkNode(ctx, *dir.Prefix, "", layouts, true, g, filterMap)
+			dirPrefix := typehelpers.SafeString(*dir.Prefix)
+			err = s.WalkNode(ctx, dirPrefix, "", layouts, true, g, filterMap)
 			if err != nil {
+				// ignore skip dir error as this means directory isn't one we want to dive into
 				if errors.Is(err, fs.SkipDir) {
 					continue
 				}
-				return fmt.Errorf("error walking node, %w", err)
+				// non-fatal error - log and notify
+				slog.Error("error obtaining directory info", "key", dirPrefix, "error", err)
+				s.NotifyError(ctx, executionId, fmt.Errorf("%s: failed to obtain directory info", dirPrefix))
+				err = nil
+				continue
 			}
-			err = s.walkS3(ctx, bucket, *dir.Prefix, layouts, filterMap, g)
+			err = s.walkS3(ctx, bucket, dirPrefix, layouts, filterMap, g)
 			if err != nil {
-				s.errorList = append(s.errorList, err)
+				// non-fatal error - log and notify
+				slog.Error("error walking S3 bucket", "bucket", bucket, "prefix", dirPrefix, "error", err)
+				s.NotifyError(ctx, executionId, fmt.Errorf("%s: %s", bucket, err.Error()))
+				err = nil
 			}
 		}
 
 		// Files
 		for _, obj := range page.Contents {
-			err = s.WalkNode(ctx, *obj.Key, "", layouts, false, g, filterMap)
+			objKey := typehelpers.SafeString(*obj.Key)
+			if objKey == "" {
+				slog.Debug("skipping empty object key")
+				continue
+			}
+			err = s.WalkNode(ctx, objKey, "", layouts, false, g, filterMap)
 			if err != nil {
-				s.errorList = append(s.errorList, fmt.Errorf("error parsing object %s, %w", *obj.Key, err))
+				// non-fatal error - log and notify
+				slog.Error("error obtaining artifact info", "key", objKey, "error", err)
+				s.NotifyError(ctx, executionId, fmt.Errorf("%s: failed to obtain artifact info", objKey))
+				err = nil
 			}
 		}
 	}