+++ tables/cost_optimization_recommendation/cost_optimization_recommendation_table.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/rs/xid"
 
-	typehelpers "github.com/turbot/go-kit/types"
 	"github.com/turbot/pipe-fittings/v2/utils"
 	"github.com/turbot/tailpipe-plugin-sdk/artifact_source"
 	"github.com/turbot/tailpipe-plugin-sdk/artifact_source_config"
@@ -35,16 +34,20 @@ func (t *CostOptimizationRecommendationsTable) GetSourceMetadata() ([]*table.Sou
 		{
 			// any artifact source
 			SourceName: constants.ArtifactSourceIdentifier,
+			Mapper:     NewCostOptimizationRecommendationMapper(),
 			Options: []row_source.RowSourceOption{
 				artifact_source.WithDefaultArtifactSourceConfig(defaultArtifactConfig),
-				artifact_source.WithArtifactExtractor(NewCostOptimizationRecommendationExtractor()),
+				artifact_source.WithRowPerLine(),
+				artifact_source.WithHeaderRowNotification(","),
 			},
 		},
 		{
 			// any artifact source
 			SourceName: constants.ArtifactSourceIdentifier,
+			Mapper:     NewCostOptimizationRecommendationMapper(),
 			Options: []row_source.RowSourceOption{
-				artifact_source.WithArtifactExtractor(NewCostOptimizationRecommendationExtractor()),
+				artifact_source.WithRowPerLine(),
+				artifact_source.WithHeaderRowNotification(","),
 			},
 		},
 	}, nil
@@ -64,11 +67,7 @@ func (t *CostOptimizationRecommendationsTable) EnrichRow(row *CostOptimizationRe
 	row.TpDate = row.LastRefreshTimestamp.Truncate(24 * time.Hour)
 
 	// TpIndex
-	if typehelpers.SafeString(row.AccountID) != "" {
-		row.TpIndex = typehelpers.SafeString(row.AccountID)
-	} else {
-		row.TpIndex = schema.DefaultIndex
-	}
+	row.TpIndex = schema.DefaultIndex
 
 	if row.ResourceARN != nil {
 		row.TpAkas = append(row.TpAkas, *row.ResourceARN)